Python 3 support
Convert the source code to support Python 3, as well as apply a few
idiomatic fixes that were spotted along the way.

For now, we keep the interpreter hashbang pointing at /usr/bin/python,
which is Python 2 on most distributions. However, Python 2 support will go
away fairly soon.

Due to a change in Python 3's int(), octal values in the configuration must
now be given as "0o123456" rather than "0123456".

Done with the help of 2to3.

Closes: #1 ("Python 3 support")
Signed-off-by: Andrew Donnellan <andrew@donnellan.id.au>